Support pattern playback. Break all existing code.
Fortunately, this library is only hours old, so the amount of existing
code should be minimal.

Colors are now passed as 3-tuples instead of separate red, green, and
blue.

Durations are now optional and default to 2 seconds.

Argument orders flipped.

You can now update the internal pattern buffer as well as start and
stop playback.