Easy pattern playback of 2 <= N <= 12 steps.
If there are fewer than 12 steps, the last one is expanded into
multiple steps to fill out the 12. This way, the user can write the
number of steps that they want without worrying about the leftover
slots stuck on the end of their pattern.

Bonus: write_pattern_line can now be called with degamma=False to pass
the color value through verbatim.