Fix message sending.
Before, it worked most of the time, but would randomly fail with a USB
pipe error. Now that seems to have cleared up. I suspect it is the
change in target from the device's interface to just the device,
though I'm very far from a USB expert, so I can't be sure.